+++ CHANGELOG.md
@@ -11,6 +11,7 @@ Placing playlist into memory causes more RAM usage but reduces disk access.
 Output filters are applied after all transformations have been performed, therefore, all filter contents must refer to the final state of the playlist.
 - Added burst buffer to shared stream
 - Telegram message thread support. thread id can now be appended to chat-id like `chat-id:thread-id`.
+- Telegram supports markdown generation for structured json messages. simply set `markdown: true` in telegram config.
 
 # 3.1.7 (2025-10-10)
 - Added Dark/Bright theme switch